agents/retriever_agent/requirements.txt
fastapi
uvicorn
faiss-cpu
langchain
sentence-transformers
